requirements.txt
# Core dependencies
python-telegram-bot==20.7
aiogram==3.4.1
requests==2.31.0
asyncio==3.4.3
typing==3.7.4.3

# AI/ML dependencies
torch==2.1.2
transformers==4.36.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3
chromadb==0.4.22

# Audio processing
sounddevice==0.4.6
soundcard==0.4.2
pydub==0.25.1
faster-whisper==0.10.0
edge-tts==6.1.10
Tortoise-TTS==3.0.0

# Computer vision
opencv-python==4.8.1.78
Pillow==10.1.0
pytesseract==0.3.10

# Desktop automation
pyautogui==0.9.54
pygetwindow==0.0.9
psutil==5.9.6

# Browser automation
playwright==1.40.0
selenium==4.16.0

# Email and calendar
icalendar==5.0.11
imapclient==2.3.1

# System utilities
python-dotenv==1.0.0
loguru==0.7.2
pywin32==306; sys_platform == 'win32'

# Development dependencies
pytest==7.4.3
black==23.12.1
flake8==6.1.0